backend/requirements.txt
# Core Python
python-dotenv==1.0.1

# FastAPI framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Web scraping
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0

# Database & ORM
SQLAlchemy==2.0.36
pymysql==1.1.1

# LangChain ecosystem
langchain-core==0.3.11
langchain-community==0.3.3
langchain-google-genai==1.0.5
sentence-transformers
faiss-cpu
langchain-huggingface

# Embeddings & Vector Store
faiss-cpu==1.8.0.post1

# Data validation
pydantic==2.9.2
pydantic-core==2.23.4

# Optional utilities
tqdm==4.66.5
numpy==1.26.4
